requirements.txt
anyio==3.4.0
astroid==2.8.6
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
dataclasses==0.6
decorator==5.1.0
deprecation==2.1.0
Flask==2.0.2
gotrue==0.2.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==2.10
iniconfig==1.1.1
ipython==7.29.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
postgrest-py==0.5.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.11.0
Pygments==2.10.0
pylint==2.11.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.2
wrapt==1.13.3
